+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/language/impl/DataFetcherImpl.java
@@ -93,7 +93,7 @@ public class DataFetcherImpl implements DataFetcher {
 
 	@Override
 	public String getName() {
-		return this.field.getCamelCaseName();
+		return this.field.getName();
 	}
 
 	/** {@inheritDoc} */