+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/generate_code/AbstractDocumentParser_Forum_Client.java
@@ -25,7 +25,7 @@ import com.graphql_java_generator.plugin.test.helper.GraphQLConfigurationTestHel
  * @author etienne-sf
  */
 @Execution(ExecutionMode.CONCURRENT)
-class AbstractDocumentParser_Forum_Client {
+abstract class AbstractDocumentParser_Forum_Client {
 
 	AbstractApplicationContext ctx = null;
 	GenerateCodeDocumentParser documentParser;