+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/generate_code/DocumentParser_allGraphQLCases_Server_Test.java
@@ -679,7 +679,7 @@ class DocumentParser_allGraphQLCases_Server_Test {
 
 		// Verification
 		assertEquals("MyQueryType", type.getName());
-		assertEquals(72, type.getFields().size());
+		assertEquals(73, type.getFields().size());
 
 		int j = 0; // The first query is 0, see ++j below