+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/impl/DataFetchersDelegateMyQueryTypeImpl.java
@@ -792,4 +792,9 @@ public class DataFetchersDelegateMyQueryTypeImpl implements DataFetchersDelegate
 			return null;
 		}
 	}
+
+	@Override
+	public String Issue217(DataFetchingEnvironment dataFetchingEnvironment, String AnArg) {
+		return AnArg;
+	}
 }